app finalizada
Faltan:

retoques de diseño (pablo)
retoques del fichero PDF -- 404 (no se porque)
falta fichero JSON -- no da tiempo

+++ routes/web.php
@@ -24,32 +24,38 @@ Route::get('/', function() {
 
 /** Rutas de pedidos Batoilogic */
 
-Route::get('/pedidos','orderController@index'); //mostrar todos los pedidos
+Route::middleware(['auth:sanctum','verified'])->get('/pedidos','orderController@index'); //mostrar todos los pedidos
 
-Route::get('/pedidos/{id}','orderController@show'); //mostrar un pedido
+Route::middleware(['auth:sanctum','verified'])->get('/pedidos/{id}','orderController@show'); //mostrar un pedido
 
-Route::get('/pedidos/crear','orderController@create'); // crear un pedido nuevo
+Route::middleware(['auth:sanctum','verified'])->get('/pedidos/crear','orderController@create'); // crear un pedido nuevo
 
-Route::get('/pedidos/{id}/editar','orderController@edit'); //editar un pedido
+Route::middleware(['auth:sanctum','verified'])->get('/pedidos/{id}/editar','orderController@edit'); //editar un pedido
 
-Route::delete('/pedidos/{id}/eliminar','orderController@destroy');//eliminar un pedido
+Route::middleware(['auth:sanctum','verified'])->delete('/pedidos/{id}/eliminar','orderController@destroy');//eliminar un pedido
 
-Route::put('/pedidos/{id}/update','orderController@update'); //efectuar la actualizacion de los datos de un pedido
+Route::middleware(['auth:sanctum','verified'])->put('/pedidos/{id}/update','orderController@update'); //efectuar la actualizacion de los datos de un pedido
+
+Route::get('/pedidos/albaran','orderController@albaran')->name('albaran'); //se descarga el pdf albaran que corresponde a las entregas del dia del repartidor
 
 
 /** Rutas de proveedores Batoilogic */
 
-Route::get('/proveedores','providerController@index'); //mostrar la lista de proveedores
+Route::middleware(['auth:sanctum','verified'])->get('/proveedores','providerController@index'); //mostrar la lista de proveedores
 
-Route::get('/proveedores/{id}','providerController@show'); //mostrar un proveedor
+Route::middleware(['auth:sanctum','verified'])->get('/proveedores/{id}','providerController@show'); //mostrar un proveedor
 
-Route::get('/proveedores/crear','providerController@create'); // crear un proveedor nuevo
+Route::middleware(['auth:sanctum','verified'])->get('/proveedores/crear','providerController@create'); // crear un proveedor nuevo
 
-Route::get('/proveedores/{id}/editar','providerController@edit'); //editar un proveedor
+Route::middleware(['auth:sanctum','verified'])->get('/proveedores/{id}/editar','providerController@edit'); //editar un proveedor
 
-Route::delete('/proveedores/{id}/eliminar','providerController@destroy');//eliminar un proveedor
+Route::middleware(['auth:sanctum','verified'])->delete('/proveedores/{id}/eliminar','providerController@destroy');//eliminar un proveedor
 
-Route::put('/proveedores/{id}/update','providerController@update'); //efectuar la actualizacion de los datos de un proveedor
+Route::middleware(['auth:sanctum','verified'])->put('/proveedores/{id}/update','providerController@update'); //efectuar la actualizacion de los datos de un proveedor
+
+Route::middleware(['auth:sanctum','verified'])->get('/proveedores/{id}/stocking','providerController@stocking'); //efectuar la actualizacion de los stocks de un proveedor
+
+Route::middleware(['auth:sanctum','verified'])->get('/proveedores/{id}/stockageUpdate','providerController@stockUpdate'); //efectuar la actualizacion de los stocks de un proveedor
 
 /** PRODUCTOS --- Batoilogic */
 
@@ -68,14 +74,14 @@ Route::put('/productos/{id}/update','productController@update'); //efectuar la a
 
 /** USUARIOS --- Batoilogic */
 
-Route::get('/usuarios','userController@index'); //mostrar lista de usuarios
+Route::middleware(['auth:sanctum','verified'])->get('/usuarios','userController@index'); //mostrar lista de usuarios
 
-Route::get('/usuario/{id}','userController@show'); //mostrar un usuario
+Route::middleware(['auth:sanctum','verified'])->get('/usuario/{id}','userController@show'); //mostrar un usuario
 
-Route::get('/usuario/crear','userController@create'); // crear un usuario nuevo
+Route::middleware(['auth:sanctum','verified'])->get('/usuario/crear','userController@create'); // crear un usuario nuevo
 
-Route::get('/usuario/{id}/editar','userController@edit'); //editar un usuario
+Route::middleware(['auth:sanctum','verified'])->get('/usuario/{id}/editar','userController@edit'); //editar un usuario
 
-Route::delete('/usuario/{id}/eliminar','userController@destroy');//eliminar un usuario
+Route::middleware(['auth:sanctum','verified'])->delete('/usuario/{id}/eliminar','userController@destroy');//eliminar un usuario
 
-Route::put('/usuario/{id}/update','userController@update'); //efectuar la actualizacion de los datos de un usuario
+Route::middleware(['auth:sanctum','verified'])->put('/usuario/{id}/update','userController@update'); //efectuar la actualizacion de los datos de un usuario
